Refactoring and fixes to the tests
- Separate complexities: formatting of the release notes with the domain object that holds the information for the release notes
- This allows us to change formatting of release notes easily
- Updated integ tests so that they refer to the plugin correctly
