+++ classes/production/mockito-release-tools/META-INF/gradle-plugins/org.mockito.release-notes.properties
@@ -1,0 +1,1 @@
+implementation-class=org.mockito.release.gradle.notes.ReleaseNotesPlugin
