+++ classes/production/mockito-release-tools/META-INF/gradle-plugins/org.mockito.release-workflow.properties
@@ -1,0 +1,1 @@
+implementation-class=org.mockito.workflow.gradle.internal.ReleaseWorkflowPlugin
