+++ src/main/groovy/org/mockito/release/notes/NotesPrinter.java
@@ -1,5 +1,6 @@
 package org.mockito.release.notes;
 
+import org.mockito.release.notes.format.DefaultFormatter;
 import org.mockito.release.notes.util.HumanReadable;
 
 import java.text.SimpleDateFormat;
@@ -8,13 +9,13 @@ import java.util.TimeZone;
 
 class NotesPrinter {
 
-    public String printNotes(String version, Date date, HumanReadable contributions, HumanReadable improvements) {
+    public String printNotes(String version, Date date, HumanReadable contributions, String improvements) {
         SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd HH:mm z");
         format.setTimeZone(TimeZone.getTimeZone("UTC"));
         String now = format.format(date);
 
         return "### " + version + " (" + now + ")" + "\n\n"
                 + contributions.toText() + "\n"
-                + improvements.toText() + "\n\n";
+                + improvements + "\n\n";
     }
 }
