+++ src/main/groovy/org/mockito/release/notes/format/DefaultFormatter.java
@@ -1,6 +1,12 @@
 package org.mockito.release.notes.format;
 
 import org.mockito.release.notes.improvements.Improvement;
+import org.mockito.release.util.MultiMap;
+
+import java.util.Collection;
+import java.util.LinkedHashSet;
+import java.util.Map;
+import java.util.Set;
 
 /**
  * Original formatter
@@ -9,5 +15,52 @@ public class DefaultFormatter {
 
     public static String format(Improvement improvement) {
         return improvement.getTitle() + " [(#" + improvement.getId() + ")](" + improvement.getUrl() + ")";
+    }
+
+    public static String format(Map<String, String> labels, Collection<Improvement> improvements) {
+        if (improvements.isEmpty()) {
+            return "* No notable improvements. See the commits for detailed changes.";
+        }
+        StringBuilder sb = new StringBuilder("* Improvements: ").append(improvements.size());
+        MultiMap<String, Improvement> byLabel = new MultiMap<String, Improvement>();
+        Set<Improvement> remainingImprovements = new LinkedHashSet<Improvement>(improvements);
+
+        //Step 1, find improvements that match input labels
+        //Iterate label first because the input labels determine the order
+        for (String label : labels.keySet()) {
+            for (Improvement i : improvements) {
+                if (i.getLabels().contains(label) && remainingImprovements.contains(i)) {
+                    remainingImprovements.remove(i);
+                    byLabel.put(label, i);
+                }
+            }
+        }
+
+        //Step 2, print out the improvements that match input labels
+        for (String label : byLabel.keySet()) {
+            String labelCaption = labels.get(label);
+            Collection<Improvement> labelImprovements = byLabel.get(label);
+            sb.append("\n  * ").append(labelCaption).append(": ").append(labelImprovements.size());
+            for (Improvement i : labelImprovements) {
+                sb.append("\n    * ").append(DefaultFormatter.format(i));
+            }
+        }
+
+        //Step 3, print out remaining changes
+        if (!remainingImprovements.isEmpty()) {
+            String indent;
+            //We want clean view depending if there are labelled improvements or not
+            if (byLabel.size() > 0) {
+                indent = "  ";
+                sb.append("\n  * Remaining changes: ").append(remainingImprovements.size());
+            } else {
+                indent = "";
+            }
+
+            for (Improvement i : remainingImprovements) {
+                sb.append("\n").append(indent).append("  * ").append(DefaultFormatter.format(i));
+            }
+        }
+        return sb.toString();
     }
 }
