+++ src/main/groovy/org/mockito/release/notes/improvements/GitHubImprovementsProvider.java
@@ -4,6 +4,7 @@ import org.mockito.release.notes.vcs.ContributionSet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.util.Collection;
 import java.util.Map;
 
 class GitHubImprovementsProvider implements ImprovementsProvider {
@@ -11,14 +12,12 @@ class GitHubImprovementsProvider implements ImprovementsProvider {
     private static final Logger LOGGER = LoggerFactory.getLogger(GitHubImprovementsProvider.class);
     private final String authToken;
 
-    public GitHubImprovementsProvider(String authToken) {
+    GitHubImprovementsProvider(String authToken) {
         this.authToken = authToken;
     }
 
-    public ImprovementSet getImprovements(ContributionSet contributions, Map<String, String> labels) {
+    public Collection<Improvement> getImprovements(ContributionSet contributions, Map<String, String> labels) {
         LOGGER.info("Parsing {} commits with {} tickets", contributions.getAllCommits().size(), contributions.getAllTickets().size());
-        DefaultImprovements out = new DefaultImprovements(labels);
-        new GitHubTicketFetcher().fetchTickets(authToken, contributions.getAllTickets(), out);
-        return out;
+        return new GitHubTicketFetcher().fetchTickets(authToken, contributions.getAllTickets());
     }
 }
