--- src/main/groovy/org/mockito/release/notes/improvements/ImprovementSet.java
@@ -1,14 +1,0 @@
-package org.mockito.release.notes.improvements;
-
-import org.mockito.release.notes.util.HumanReadable;
-
-/**
- * the improvements
- */
-public interface ImprovementSet extends HumanReadable {
-
-    /**
-     * human readable representation of the improvements
-     */
-    String toText();
-}
