+++ src/main/groovy/org/mockito/release/notes/improvements/ImprovementsProvider.java
@@ -2,9 +2,10 @@ package org.mockito.release.notes.improvements;
 
 import org.mockito.release.notes.vcs.ContributionSet;
 
+import java.util.Collection;
 import java.util.Map;
 
 public interface ImprovementsProvider {
 
-    ImprovementSet getImprovements(ContributionSet contributions, Map<String, String> labels);
+    Collection<Improvement> getImprovements(ContributionSet contributions, Map<String, String> labels);
 }
