+++ src/test/groovy/org/mockito/release/notes/improvements/GitHubTicketFetcherTest.groovy
@@ -1,5 +1,6 @@
 package org.mockito.release.notes.improvements
 
+import org.mockito.release.notes.format.DefaultFormatter
 import spock.lang.Ignore
 import spock.lang.Specification
 import spock.lang.Subject
@@ -15,12 +16,13 @@ class GitHubTicketFetcherTest extends Specification {
     //It's not ideal but it gives us a good smoke test
     //So far it is not problematic to maintain :)
     def "fetches improvements from GitHub"() {
-        def impr = new DefaultImprovements([:])
         def readOnlyToken = "a0a4c0f41c200f7c653323014d6a72a127764e17"
-        when: fetcher.fetchTickets(readOnlyToken, ['109', '108', '99999', '112'], impr)
+        when:
+        def improvements = fetcher.fetchTickets(readOnlyToken, ['109', '108', '99999', '112']) as List
+
         then:
-        impr.improvements.get(0).labels == ["enhancement"] as Set
-        impr.toText() == """* Improvements: 3
+        improvements[0].labels == ["enhancement"] as Set
+        DefaultFormatter.format([:], improvements) == """* Improvements: 3
   * Allow instances of other classes in AdditionalAnswers.delegatesTo [(#112)](https://github.com/mockito/mockito/issues/112)
   * Improve automated release notes look [(#109)](https://github.com/mockito/mockito/issues/109)
   * Clarify Spy vs Mock CALLS_REAL_METHODS [(#108)](https://github.com/mockito/mockito/issues/108)"""
